### 2022-01-11 #498 @clnsmth
EDIutils
**Reviewers:**

Submitting Author Name: Colin Smith
Submitting Author Github Handle: @clnsmth
Repository:  https://github.com/EDIorg/EDIutils
Version submitted:
Submission type: Standard
Version accepted: TBD

---

```
Package: EDIutils
Title: An API Client for the Environmental Data Initiative Repository
Version: 0.0.0.9000
    c(person("Colin", "Smith", email = "colin.smith@wisc.edu", role = c("aut", "cre"), comment = "https://orcid.org/0000-0003-2261-9931"),
    person("Corinna", "Gries", role = "ctb", comment = "0000-0002-9091-6543"))
Description: A client for the Environmental Data Initiative repository REST API. The EDI data repository <https://portal.edirepository.org/nis/home.jsp> is for publication and reuse of ecological data with emphasis on metadata accuracy and completeness. It is built upon the PASTA+ software stack <https://pastaplus-core.readthedocs.io/en/latest/index.html#> and was developed in collaboration with the US LTER Network <https://lternet.edu/>. EDIutils includes functions to search and access existing data, evaluate and upload new data, and assist other data management tasks common to repository users.
    curl,
    httr,
    xml2
    roxygen2,

```

	- [x ] data retrieval
	- [ x] data deposition

The package is an API client for the Environmental Data Initiative data repository, which serves the ecological and environmental science and monitoring domains.

The target audiences are:
1.) Data authors who publish to the repository.
2.) Data users who retrieve data from the repository.

There is no overlap.

Yes

All checks pass

- [x ] I have read the [guide for authors](https://devguide.ropensci.org/guide-for-authors.html) and [rOpenSci packaging guide](https://devguide.ropensci.org/building.html).

- [ x] does not violate the Terms of Service of any service it interacts with.
- [ x] has a CRAN and OSI accepted license.
- [ x] contains a [README with instructions for installing the development version](https://ropensci.github.io/dev_guide/building.html#readme).
- [ x] includes [documentation with examples for all functions, created with roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation).
- [ x] contains a vignette with examples of its essential functions and uses.
- [ x] has a [test suite](https://ropensci.github.io/dev_guide/building.html#testing).
- [ x] has [continuous integration](https://ropensci.github.io/dev_guide/ci.html), including reporting of test coverage using services such as Travis CI, Coveralls and/or CodeCov.

- [ x] Do you intend for this package to go on CRAN?

- [ x] I agree to abide by [rOpenSci's Code of Conduct](https://ropensci.org/code-of-conduct/) during the review process and in maintaining my package should it be accepted.
