### 2022-01-14 #498 @ldecicco-USGS
Hi @clnsmth ! Thanks for the submission. The scope of the package looks like a good fit for rOpenSci. I'm going to assign an editor.

As for the test coverage, have a look here:
https://books.ropensci.org/http-testing/
And see if there are some techniques you could use to bump up the coverage.

Thanks for the submission!
