### 2022-01-14 #498 @clnsmth
Hi @ldecicco-USGS and @adamhsparks ! Thanks for the http-testing resource. I'll send an update once the requested changes are implemented.
