### 2022-01-20 032d42b
Implement rOpenSci http testing recommendations
Based on https://books.ropensci.org/http-testing/index.html
- Use vcr to mock http requests
- Run tests with real http requests via GitHub Actions on a monthly
  schedule and on push and pull requests to the main branch
- Update code coverage badge on the main branch when tests use real
  http requests
- Precompute all examples and vignettes
