+++ tests/testthat/test_list_deleted_data_packages.R
@@ -1,7 +1,9 @@
 context("List deleted data packages")
 
-testthat::test_that("Test attributes of returned object", {
-  res <- list_deleted_data_packages()
+testthat::test_that("list_deleted_data_packages() works", {
+  vcr::use_cassette("list_deleted_data_packages", {
+    res <- list_deleted_data_packages()
+  })
   expect_equal(class(res), "character")
   expect_true(length(res) > 0)
 })
