+++ tests/testthat/test_read_data_entity_checksum.R
@@ -1,9 +1,11 @@
 context("Read data entity checksum")
 
-testthat::test_that("Test attributes of returned object", {
+testthat::test_that("read_data_entity_checksum() works", {
   packageId <- "knb-lter-ble.1.7"
-  entityId <- list_data_entities(packageId)
-  res <- read_data_entity_checksum(packageId, entityId[1])
+  entityId <- "a1723e0e5f3c4881f1a7ede1b036aba6"
+  vcr::use_cassette("read_data_entity_checksum", {
+    res <- read_data_entity_checksum(packageId, entityId)
+  })
   expect_equal(class(res), "character")
   expect_true(length(res) > 0)
 })
