+++ tests/testthat/test_evaluate_report_summary.R
@@ -1,5 +1,9 @@
 context("Summarize evaluate report")
 
-testthat::test_that("Test attributes of returned object", {
-  # Tested in test_check_status_evaluate.R
+testthat::test_that("evaluate_report_summary() works", {
+  transaction <- "evaluate_163966785813042760"
+  vcr::use_cassette("read_evaluate_report_summary", {
+    expect_message(
+      read_evaluate_report_summary(transaction, env = "staging"))
+  })
 })
