+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 2.7.3
-pkgdown: 2.0.1
+pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   evaluate_and_upload: evaluate_and_upload.html
@@ -7,5 +7,5 @@ articles:
   retrieve_downloads: retrieve_downloads.html
   search_and_access: search_and_access.html
   tests_requiring_authentication: tests_requiring_authentication.html
-last_built: 2022-01-11T21:30Z
+last_built: 2022-01-20T21:38Z
 
