+++ .github/workflows/test-coverage.yaml
@@ -1,35 +1,0 @@
-# Workflow derived from https://github.com/r-lib/actions/tree/master/examples
-# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
-
-name: test-coverage
-
-on:
-  push:
-    branches: [main, development]
-  pull_request:
-    branches: [main, development]
-
-jobs:
-  test-coverage:
-    runs-on: ubuntu-latest
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      EDI_USER: ${{ secrets.EDI_USER }}
-      EDI_PASS: ${{ secrets.EDI_PASS }}
-      CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN}}
-      VCR_TURN_OFF: true
-
-    steps:
-      - uses: actions/checkout@v2
-
-      - uses: r-lib/actions/setup-r@v1
-        with:
-          use-public-rspm: true
-
-      - uses: r-lib/actions/setup-r-dependencies@v1
-        with:
-          extra-packages: covr
-
-      - name: Test coverage
-        run: covr::codecov()
-        shell: Rscript {0}
