+++ .github/workflows/check-standard-real-requests.yaml
@@ -62,3 +62,28 @@ jobs:
         with:
           name: ${{ runner.os }}-r${{ matrix.config.r }}-results
           path: check
+  
+  test-coverage:
+    needs: R-CMD-check
+    runs-on: ubuntu-latest
+    env:
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      EDI_USER: ${{ secrets.EDI_USER }}
+      EDI_PASS: ${{ secrets.EDI_PASS }}
+      CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+      VCR_TURN_OFF: true
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - uses: r-lib/actions/setup-r@v1
+        with:
+          use-public-rspm: true
+
+      - uses: r-lib/actions/setup-r-dependencies@v1
+        with:
+          extra-packages: covr
+
+      - name: Test coverage
+        run: covr::codecov()
+        shell: Rscript {0}
