### 2022-01-21 #498 @clnsmth
Hi @adamhsparks @ldecicco-USGS,

I've implement rOpenSci http testing recommendations and coverage is now > 75%.

EDIutils:
- Uses vcr to mock http requests
- Runs tests with real http requests via GitHub Actions on a monthly schedule and on push and pull requests to the main branch
- Precomputes all examples and vignettes

Thanks again for this helpful advice and please let me know if I've missed anything.
