### 2022-01-22 6d43c49
Wrap example in \dontrun{}
Some how this was missed.
