### 2022-02-14 #498 @laijasmine
---

Thanks for the opportunity to be involved in this! It was interesting to see how other data repositories deal with data upload. I also wanted to preface this by saying my experience is heavily influenced by data repository access and submission package, `dataone` and working on editing EML using the `EML` package. The workflow there is very similar but different in some elements so some of the things I might not be applicable to your situation!

##### Search and Access
Many users are probably unfamiliar with xml and how to navigate it and there is a learning curve involved so a vignette on the files would be helpful. They might be more comfortable with working with dataframes instead. We don't need to go into detail but it might be helpful to point them to packages like XML to help convert the content to a preferred format. Maybe adding a couple lines or point them to retrieve downloads will be helpful.

##### Evaluate and upload
*Upload a data package*
I struggled the longest here. Maybe because I was just trying to write the least amount of information possible to just upload the file. I had a hard time understanding what the errors were in the results of `read_evaluate_report`. I went back and forth between the ezEML form and R trying to figure out what I needed. There seems to be more requirements to upload using `EDIutils` compared to the Check Metadata in ezEML. For example I tried to avoid uploading a data table because it wasn't marked as required in ezEML but it was required to upload using `EDIutils`.

*Report*
I also found the output from `read_evaluate_report`quite long and hard to read if I'm viewing it as a message in the console. We might want to suggest users to use `writeLines(report, "report.txt")` instead so they can cmd + f the file and keep track of their fixes. Users might prefer if there is a way to only show the warnings and errors so there is less to look through when trying to improve their metadata.

*Update a data package*
Following the example in the vignette I wasn't able to update my package because I keep getting the error:
```
Error: Attempting to update a data package to revision '1' but an equal or higher revision ('1') already exists in PASTA: edi.757.1
```
I realized the `packageId` slot inside the eml file also needs to be updated. We might need to be more explicit in the vignette that the identifier needs to be updated. I used the following EML package here:
```
my_eml <- EML::read_eml("edi.757.1.xml")
my_eml$dataset$title <- "test 2 update the title :)"
my_eml$packageId <- "edi.757.2"
EML::write_eml(my_eml, "edi.757.2.xml")
```

##### list_principal_owner_citations
in the documentation it mentions the principalOwner as being:
> (character) Principal owner in the format returned by construct_dn()

I can't seem to find that function. I think this should say `create_dn()`?

##### search_data_package
It might be helpful for the documentation to have the formats for some of the fields such as singledate, begindate and pubdate so it can be easily referenced.

##### create_journal_citation
using the example (and delete_journal_citation example) and got the following error:
```
journalCitationId <- create_journal_citation(
  packageId = "edi.17.1",
  articleDoi = "https://doi.org/10.1890/11-1026.1",
  articleTitle = "Corridors promote fire via connectivity and edge effects",
  journalTitle = "Ecological Applications",
  relationType = "IsCitedBy",
  env = "staging"
)
# Error in create_journal_citation(packageId = "edi.17.1", articleDoi = "https://doi.org/10.1890/11-1026.1",  :
#   Bad Request (HTTP 400). Failed to DOI https://doi.org/10.1890/11-1026.1 should be in "prefix/suffix" format (e.g., 10.6073/d4f26a6c).
```
I think the articleDOI in the example should be "10.1890/11-1026.1".
