### 2022-02-23 #498 @bozaah
**No current or past work/personal relationship to report.**

Estimated hours spent reviewing: ~9h

---

Thank you for the opportunity to review the `EDIutils` package. This package provides an API client to access the Environmental Data Initiative Repository, in addition to the web browser portal ([EDI](https://environmentaldatainitiative.org)). It consists of a suite of functions to browse/list, access, evaluate and upload data packages as well as manage citations entries and usage metrics associated with the data packages.

I am reviewing the package using *R version 4.1.2 (2021-11-01)* and platform *x86_64-apple-darwin17.0 (64-bit)*. No issues reported when running `devtools::check()`. When running `goodpractice::gp()` no errors/issues however *unit testing* warns that *77% of code lines are covered by test cases*.

Overall, vignettes are well structured, complete and cover the general use and applicability of the package functions.

##### Search and Access
No issues with the search and access vignettes. I also mirror comments from @laijasmine regarding XML which, I am not too familiar with and, it is likely the reality of many potential users of the package. There is already a note on *Retrieve Download Metrics* section which mention that several functions of the EDI API return XML and that `xml2` is a reasonable approach to locally manipulate (eg extract and save) EML files.

Several functions do not work until the user has completed the authentication step with `login()` which requires a EDI account. The vignette *Tests Requiring Authentication* has information about it but perhaps a vignette on authentication only could be a reasonable addition. If so, I believe it should placed as the first vignette or together with a *Get started* vignette.

Lastly, mirroring comments of @laijasmine regarding the doi formats, the function `read_data_package_doi()` returns only the doi identifier eg doi:10.6073/pasta/5d2c5ac4d6cb1293a47cfa6ef5c55a01 but it could perhaps return the complete URL ie https://doi.org/10.6073/pasta/5d2c5ac4d6cb1293a47cfa6ef5c55a01

##### Evaluate and Upload
No issues here but I also agree with comments about the output of the `read_evalute_report()` function. However, the function description/example does present @laijasmine's suggestion to use `writeLines()` and also mention the possibility of alternative output formats with the `frmt` argument (defaults to xml). I also noted that the `read_evalute_report_summary()` offers a compact way to visualise the report output, which is quite neat.

##### Retrieve Download and Citation metrics
These vignettes are well structure and include reasonable examples and potential user cases. Functions and arguments are descriptive and detailed. No errors or issues to report for those vignettes, with exception of the citation doi error in the examples which were already mentioned by @laijasmine.

##### Documentation and examples
Documentation is automatically generated with `roxygen2` and `devtools::document()`. Functionally is descriptive and complete. R code follows community guidelines (ie tidyverse style guide and rOpenSci package guide). The author also included contribution guidelines separately from the README. Examples run without errors.

#### Final Remarks
**Access to data repositories with accurate and structured/complete metadata is essential for research reproducibility and general data reuse.  The ability to obtain usage and citation metrics for the data packages in the EDI repository is outstading. `EDIutils` offers a great tool for the Open Source community and it will be a welcomed addition to the rOpenSci packages.**
