+++ man/read_data_package_archive.Rd
@@ -52,6 +52,7 @@ Other Accessing:
 \code{\link{read_data_entity_sizes}()},
 \code{\link{read_data_entity_size}()},
 \code{\link{read_data_entity}()},
+\code{\link{read_data_package_citation}()},
 \code{\link{read_data_package_doi}()},
 \code{\link{read_data_package_error}()},
 \code{\link{read_data_package_from_doi}()},
