+++ NAMESPACE
@@ -53,6 +53,7 @@ export(read_data_entity_size)
 export(read_data_entity_sizes)
 export(read_data_package)
 export(read_data_package_archive)
+export(read_data_package_citation)
 export(read_data_package_doi)
 export(read_data_package_error)
 export(read_data_package_from_doi)
