### 2022-03-04 61452cc
updated download metrics vignette.
