### 2022-03-04 #498 @clnsmth
Thanks again for these helpful reviews @laijasmine @bozaah! I agree with all of your recommendations and will begin implementing them in the `development` branch of the EDIutils GitHub. The responses and links to corresponding commits will be listed here in this issue thread, and the changes should be completed by the end of Tuesday.
