### 2022-03-07 f38a1fd
Update audit report and record examples
The create and get audit record methods now return the userAgent in results. They were previously returning NULL values.
