### 2022-03-07 082b2ab
Reference correct function
A non-existant function construct_dn() was being referenced where it should have been create_dn().
