### 2022-03-07 743785c
Fix DOI format
The example for create_journal_citation() was using the incorrect format.
