### 2022-03-08 82f38bc
Add xml2df()
A non-exported function for converting basic xml to a data.frame. This function will allow users to get returned results in an "easier to use" data.frame format.
