+++ tests/testthat/test_utilities.R
@@ -83,3 +83,18 @@ testthat::test_that("text2char() works", {
   res <- text2char("text\ntext\n")
   expect_length(res, 2)
 })
+
+
+
+testthat::test_that("xml2df() works", {
+  # One level of nesting
+  resourceMetadata <- read_data_package_report_resource_metadata(
+    packageId = "knb-lter-mcm.9129.3"
+  )
+  res <- xml2df(resourceMetadata)
+  expect_equal(class(res), "data.frame")
+  # Two levels of nesting
+  dataDescendants <- list_data_descendants("knb-lter-bnz.501.17")
+  res <- xml2df(dataDescendants)
+  expect_equal(class(res), "data.frame")
+})
