### 2022-03-08 0441ba2
Apply xml2df() to exported functions
