+++ R/get_audit_report.R
@@ -1,10 +1,12 @@
 #' Get audit report
 #'
 #' @param query (character) Query (see details below)
+#' @param as (character) Format of the returned object. Can be: "data.frame" 
+#' or "xml".
 #' @param env (character) Repository environment. Can be: "production",
 #' "staging", or "development".
 #'
-#' @return (xml_document) An XML list of zero or more audit records matching
+#' @return (data.frame or xml_document) Zero or more audit records matching
 #' the query parameters as specified in the request (see details below).
 #'
 #' @details Query parameters are specified as key=value pairs, multiple pairs
@@ -55,39 +57,16 @@
 #' # Get audit report for data reads between 2021-12-01 and 2021-12-02
 #' query <- "serviceMethod=readDataEntity&fromTime=2021-12-01&toTime=2021-12-02"
 #' auditReport <- get_audit_report(query)
-#' auditReport
-#' #> [1] <auditRecord>\n  <oid>121606334</oid>\n  <entryTime>2021-12-01T ...
-#' #> [2] <auditRecord>\n  <oid>121606349</oid>\n  <entryTime>2021-12-01T ...
-#' #> [3] <auditRecord>\n  <oid>121606465</oid>\n  <entryTime>2021-12-01T ...
-#' #> [4] <auditRecord>\n  <oid>121606937</oid>\n  <entryTime>2021-12-01T ...
-#' #> [5] <auditRecord>\n  <oid>121607211</oid>\n  <entryTime>2021-12-01T ...
-#' #> ...
-#'
-#' # Get the first audit record
-#' xml2::xml_find_first(auditReport, ".//auditRecord")
-#' #> {xml_node}
-#' #> <auditRecord>
-#' #>  [1] <oid>121606334</oid>
-#' #>  [2] <entryTime>2021-12-01T00:00:07</entryTime>
-#' #>  [3] <category>warn</category>
-#' #>  [4] <service>DataPackageManager-1.0</service>
-#' #>  [5] <serviceMethod>readDataEntity</serviceMethod>
-#' #>  [6] <responseStatus>401</responseStatus>
-#' #>  [7] <resourceId/>
-#' #>  [8] <user>robot</user>
-#' #>  [9] <userAgent>Mozilla/5.0 (compatible; SemrushBot/7~bl; +http://ww ...
-#' #> [10] <groups/>
-#' #> [11] <authSystem>https://pasta.edirepository.org/authentication</aut ...
-#' #> [12] <entryText>Robots are not authorized access to data objects. Ro ...
 #'
 #' logout()
 #' }
 #'
-get_audit_report <- function(query, env = "production") {
+get_audit_report <- function(query, as = "data.frame", env = "production") {
   url <- paste0(base_url(env), "/audit/report?", query)
   cookie <- bake_cookie()
   resp <- httr::GET(url, set_user_agent(), cookie, handle = httr::handle(""))
   res <- httr::content(resp, as = "text", encoding = "UTF-8")
   httr::stop_for_status(resp, res)
-  return(xml2::read_xml(res))
+  res <- xml2::read_xml(res)
+  ifelse(as == "data.frame", return(xml2df(res)), return(res))
 }
