+++ tests/testthat/test_list_recent_uploads.R
@@ -2,7 +2,7 @@ context("List recent uploads")
 
 testthat::test_that("list_recent_uploads() works", {
   vcr::use_cassette("list_recent_uploads", {
-    res <- list_recent_uploads(type = "update", 5)
+    res <- list_recent_uploads(type = "update", limit = 5, as = "xml")
   })
   expect_true(all(class(res) %in% c("xml_document", "xml_node")))
   expect_true("dataPackage" %in% xml2::xml_name(xml2::xml_children(res)))
