### 2022-03-09 26f888a
Note where to get help working with complex XML
Not all functions can return XML in a simplified format, so point users to some resources.
