+++ R/read_metadata.R
@@ -4,7 +4,13 @@
 #' @param env (character) Repository environment. Can be: "production",
 #' "staging", or "development".
 #'
-#' @return (xml_document) EML metadata document
+#' @return (xml_document) EML metadata document.
+#' 
+#' See the 
+#' \href{https://cran.r-project.org/web/packages/emld/index.html}{emld} library 
+#' for more on working with EML as a list or JSON-LD. See the 
+#' \href{https://cran.r-project.org/web/packages/xml2/index.html}{xml2} library 
+#' for working with EML as XML.
 #' 
 #' @family Accessing
 #'
