+++ R/read_metadata_dublin_core.R
@@ -4,7 +4,11 @@
 #' @param env (character) Repository environment. Can be: "production",
 #' "staging", or "development".
 #'
-#' @return (xml_document) Dublin Core metadata
+#' @return (xml_document) Dublin Core metadata.
+#' 
+#' See the 
+#' \href{https://cran.r-project.org/web/packages/xml2/index.html}{xml2} library 
+#' for more on working with XML.
 #' 
 #' @family Accessing
 #'
