### 2022-03-09 EDIorg/EDIutils#34 @clnsmth
Create vignette on working with EML and XML
A great recommendation from @laijasmine and @bozaah to help get users working with EML and XML.
