### 2022-03-09 EDIorg/EDIutils#35 @clnsmth
Add parameter to report functions for only listing warns and errors
A recommendation from @laijasmine, but not yet implemented.
