+++ R/read_evaluate_report.R
@@ -28,6 +28,22 @@
 #' transaction
 #' #> [1] "evaluate_163966785813042760"
 #'
+#' # Read as HTML and write to file for a web browser view
+#' qualityReport <- read_evaluate_report(
+#'   transaction = transaction,
+#'   as = "html",
+#'   env = "staging"
+#' )
+#' writeLines(qualityReport, "./data/report.html")
+#'
+#' # Read as character and write to file for browsing
+#' qualityReport <- read_evaluate_report(
+#'   transaction = transaction,
+#'   as = "char",
+#'   env = "staging"
+#' )
+#' writeLines(qualityReport, "./data/report.txt")
+#' 
 #' # Read as XML
 #' qualityReport <- read_evaluate_report(
 #'   transaction = transaction,
@@ -42,25 +58,6 @@
 #' #> [4] <includeSystem>knb</includeSystem>
 #' #> [5] <datasetReport>\n  <qualityCheck qualityType="metadata" system=" ...
 #' #> [6] <entityReport>\n  <entityName>data.txt</entityName>\n  <qualityC ...
-#'
-#' # Read as HTML
-#' qualityReport <- read_evaluate_report(
-#'   transaction = transaction,
-#'   as = "html",
-#'   env = "staging"
-#' )
-#' qualityReport
-#' #> {html_document}
-#' #> <html>
-#' #> [1] <body><table xmlns:qr="eml://ecoinformatics.org/qualityReport">< ...
-#'
-#' # Read as character
-#' qualityReport <- read_evaluate_report(
-#'   transaction = transaction,
-#'   as = "char",
-#'   env = "staging"
-#' )
-#' writeLines(qualityReport, "./data/report.txt")
 #'
 #' logout()
 #' }
