### 2022-03-09 ef4b691
Add formats required by search fields
