### 2022-03-09 EDIorg/EDIutils#36 @clnsmth
Demonstrate common search patterns using Solr queries
Add these to the "Search and Access" vignette.
