### 2022-03-09 d34018c
Add option to get DOI as URL
