### 2022-03-09 bd49970
Highlight authentication requirements
