+++ vignettes/search_and_access.Rmd
@@ -37,19 +37,6 @@ Results can be filtered to include only fields of interest.
 ```{r eval=FALSE}
 # Match all documents with keywords "disturbance" and return only their IDs
 res <- search_data_packages(query = 'q=keyword:disturbance&fl=id')
-res
-#> {xml_document}
-#> <resultset numFound="12845" start="0" rows="10">
-#>   [1] <document>\n  <id>knb-lter-luq.197</id>\n</document>
-#>   [2] <document>\n  <id>knb-lter-and.3982</id>\n</document>
-#>   [3] <document>\n  <id>knb-lter-luq.169</id>\n</document>
-#>   [4] <document>\n  <id>knb-lter-and.3983</id>\n</document>
-#>   [5] <document>\n  <id>knb-lter-arc.10435</id>\n</document>
-#>   [6] <document>\n  <id>knb-lter-arc.10436</id>\n</document>
-#>   [7] <document>\n  <id>knb-lter-arc.10450</id>\n</document>
-#>   [8] <document>\n  <id>knb-lter-arc.10451</id>\n</document>
-#>   [9] <document>\n  <id>knb-lter-arc.10452</id>\n</document>
-#>   [10] <document>\n  <id>knb-lter-arc.10453</id>\n</document>
 ```
 
 When constructing a query note that the 15403 data packages of the [ecotrends](https://lternet.edu/the-ecotrends-project/) project and 10492 data packages of the [LTER Landsat](https://lternet.edu/lter-remote-sensing-and-geographic-information-system-data/) project, can be excluded from the returned results by including `&fq=-scope:(ecotrends+lter-landsat)` in the query string.
@@ -59,19 +46,6 @@ When constructing a query note that the 15403 data packages of the [ecotrends](h
 # lter-landsat scopes from the returned results
 query <- 'q=keyword:disturbance&fl=packageid&fq=-scope:(ecotrends+lter-landsat)'
 res <- search_data_packages(query)
-res
-#> {xml_document}
-#> <resultset numFound="12841" start="0" rows="10">
-#>   [1] <document>\n  <packageid>knb-lter-luq.197.62</packageid>\n</document>
-#>   [2] <document>\n  <packageid>knb-lter-and.3982.7</packageid>\n</document>
-#>   [3] <document>\n  <packageid>knb-lter-luq.169.65</packageid>\n</document>
-#>   [4] <document>\n  <packageid>knb-lter-and.3983.4</packageid>\n</document>
-#>   [5] <document>\n  <packageid>knb-lter-arc.10435.4</packageid>\n</document>
-#>   [6] <document>\n  <packageid>knb-lter-arc.10436.4</packageid>\n</document>
-#>   [7] <document>\n  <packageid>knb-lter-arc.10450.4</packageid>\n</document>
-#>   [8] <document>\n  <packageid>knb-lter-arc.10451.4</packageid>\n</document>
-#>   [9] <document>\n  <packageid>knb-lter-arc.10452.4</packageid>\n</document>
-#>   [10] <document>\n  <packageid>knb-lter-arc.10453.4</packageid>\n</document>
 ```
 
 Use wild cards operators to match anything and control the number of returned results.
@@ -79,10 +53,6 @@ Use wild cards operators to match anything and control the number of returned re
 ```{r eval=FALSE}
 # Match anything, display all fields, limit to only one document
 res <- search_data_packages(query = 'q=*&fl=*&rows=1')
-res
-#> {xml_document}
-#> <resultset numFound="44907" start="0" rows="1">
-#>   [1] <document>\n  <abstract>The EcoTrends project was established in ...
 ```
 
[…]
