### 2022-03-09 #498 @clnsmth
Many thanks @laijasmine and @bozaah for your comments and suggested improvments to EDIutils! I've implemented your recommendations and responded to your comments below. Additional comments and suggestions are much appreciated.

Changes to EDIutils since your review:
* Added support for the EDI "cite" service. This provides a way for users to get a citation of a data package in the EDI repository in a community recommended format. See EDIorg/EDIutils@25e1da7.
* Update audit report and record examples. The create and get audit record methods now return the userAgent in results. They were previously returning NULL values. This was a change in the underlying EDI Data Repository Software.

## Response to @laijasmine

You have been added to the EDIutils package DESCRIPTION as a reviewer (see EDIorg/EDIutils@9f3ace8). Please check the accuracy of this information and let me know if anything should be added/changed.

### Vignettes
#### Search and Access
>Many users are probably unfamiliar with xml and how to navigate it and there is a learning curve involved so a vignette on the files would be helpful. They might be more comfortable with working with dataframes instead. We don't need to go into detail but it might be helpful to point them to packages like XML to help convert the content to a preferred format. Maybe adding a couple lines or point them to retrieve downloads will be helpful.

I totally agree. The learning curve involved with XML can be off-putting and an unnecessary barrier.

Functions that returned a simple XML structure have a new `as` parameter through which the return type can be controlled with `data.frame` as the default (see EDIorg/EDIutils@82f38bc and EDIorg/EDIutils@0441ba2). Unfortunately these simplifications can't be made for more complex returns (e.g. EML) so in these cases a note has been added to the associated function metadata and vignettes pointing users to other resources for more on parsing (see EDIorg/EDIutils@26f888a).

The recommendation for a vignette on working with more complex XML is definitely well recieved, and I would be happy to create it if given a little more time. Until then, I've opened an issue in the project GitHub for this recommendation (see https://github.com/EDIorg/EDIutils/issues/34).

#### Evaluate and upload
##### Upload a data package
>I struggled the longest here. Maybe because I was just trying to write the least amount of information possible to just upload the file. I had a hard time understanding what the errors were in the results of read_evaluate_report. I went back and forth between the ezEML form and R trying to figure out what I needed. There seems to be more requirements to upload using EDIutils compared to the Check Metadata in ezEML. For example I tried to avoid uploading a data table because it wasn't marked as required in ezEML but it was required to upload using EDIutils.

My apologies for the troubles. This is clearly a short comming in the documentation and possibly the report itself.

A new section on "Interpreting the Evaluation Report" has been added to the **Evaluate and Upload Data** vignette (see EDIorg/EDIutils@f4a0ab9). This provides more information on the report structure and content. Ultimately the report itself should be updated, it's a bit cryptic in places, but that is a task that will have to be addressed at the repository level.

##### Report
>I also found the output from read_evaluate_report quite long and hard to read if I'm viewing it as a message in the console. We might want to suggest users to use writeLines(report, "report.txt") instead so they can cmd + f the file and keep track of their fixes. Users might prefer if there is a way to only show the warnings and errors so there is less to look through when trying to improve their metadata.

The suggestion for writing the report to file (for ease of viewing) was listed in the original examples of `read_evaluate_report()` but admittingly burried too deep to be noticed. Ooops. I've reordered the examples with the preferred approaches listed first (see EDIorg/EDIutils@6aaa552).

A suggestion to only return the warnings and errors would make a nice enhancement. I'll implement this if given a little more time. I've opened an issue in the project GitHub so this isn't lost (see https://github.com/EDIorg/EDIutils/issues/35).

##### Update a data package
>Following the example in the vignette I wasn't able to update my package because I keep getting the error:
>```
>Error: Attempting to update a data package to revision '1' but an equal or higher revision ('1') already exists in PASTA: edi.757.1
>```
>I realized the packageId slot inside the eml file also needs to be updated. We might need to be more explicit in the vignette that the identifier needs to be updated. I used the following EML package here:
>```
>my_eml <- EML::read_eml("edi.757.1.xml")
>my_eml$dataset$title <- "test 2 update the title :)"
>my_eml$packageId <- "edi.757.2"
>EML::write_eml(my_eml, "edi.757.2.xml")
>```

Yes, definitely a nuance worth highlighting. I've added a note to the "Update" section of the "Evaluate and Upload" vignette. Additionally, I've elevated the "Update" section level to make visible in the vignette table of contents. See EDIorg/EDIutils@c1647e0.

### Documentation
#### list_principal_owner_citations
>in the documentation it mentions the principalOwner as being:
>(character) Principal owner in the format returned by construct_dn()
>I can't seem to find that function. I think this should say create_dn()?

Thanks for catching this. Fixed in EDIorg/EDIutils@082b2ab

#### search_data_package
> It might be helpful for the documentation to have the formats for some of the fields such as singledate, begindate and pubdate so it can be easily referenced.

I completely agree and have added formats to the fields requiring one (see EDIorg/EDIutils@ef4b691). Relatedly, I've opened an issue requesting more examples demonstrating common search patterns and Solr syntax (see https://github.com/EDIorg/EDIutils/issues/36).

#### create_journal_citation
>using the example (and delete_journal_citation example) and got the following error:
>```
>journalCitationId <- create_journal_citation(
>  packageId = "edi.17.1",
>  articleDoi = "https://doi.org/10.1890/11-1026.1",
>  articleTitle = "Corridors promote fire via connectivity and edge effects",
>  journalTitle = "Ecological Applications",
>  relationType = "IsCitedBy",
>  env = "staging"
>)
> # Error in create_journal_citation(packageId = "edi.17.1", articleDoi = "https://doi.org/10.1890/11-1026.1",  :
> #   Bad Request (HTTP 400). Failed to DOI https://doi.org/10.1890/11-1026.1 should be in "prefix/suffix" format (e.g., 10.6073/d4f26a6c).
>```

Thanks for finding this. Fixed in EDIorg/EDIutils@743785c

## Response to @bozaah

You have been added to the EDIutils package DESCRIPTION as a reviewer (see EDIorg/EDIutils@9f3ace8). Please check the accuracy of this information and let me know if anything should be added/changed.

### General Comments
>I am reviewing the package using R version 4.1.2 (2021-11-01) and platform x86_64-apple-darwin17.0 (64-bit). No issues reported when running devtools::check(). When running goodpractice::gp() no errors/issues however unit testing warns that 77% of code lines are covered by test cases.

Yes, this catches my eye too. 77% isn't terribly impressive, but it bumps up to 84% when authenticated, and higher when testing data package evalate, upload, and update operations). I'll stay on the look out for ways of improving the test suite.

### Vignettes
#### Search and Access
>No issues with the search and access vignettes. I also mirror comments from @laijasmine regarding XML which, I am not too familiar with and, it is likely the reality of many potential users of the package. There is already a note on Retrieve Download Metrics section which mention that several functions of the EDI API return XML and that xml2 is a reasonable approach to locally manipulate (eg extract and save) EML files.

Agreed. XML isn't much fun for infrequent users. Functions now have the option to return a data.frame instead of xml (in most cases). For details see my response to @laijasmine above.

>Several functions do not work until the user has completed the authentication step with login() which requires a EDI account. The vignette Tests Requiring Authentication has information about it but perhaps a vignette on authentication only could be a reasonable addition. If so, I believe it should placed as the first vignette or together with a Get started vignette.

Others have requested help on this topic as well. It must need it. : )

A general statement about authentication and which types of functions require it, is now high up in the README (see EDIorg/EDIutils@bd49970). I'd be happy to write up a vignette on authentication if you think more is needed.

>Lastly, mirroring comments of @laijasmine regarding the doi formats, the function read_data_package_doi() returns only the doi identifier eg doi:10.6073/pasta/5d2c5ac4d6cb1293a47cfa6ef5c55a01 but it could perhaps return the complete URL ie https://doi.org/10.6073/pasta/5d2c5ac4d6cb1293a47cfa6ef5c55a01

Nice! This function now returns the complete URL through the `as_url` parameter (see EDIorg/EDIutils@d34018c).

#### Evaluate and Upload
>No issues here but I also agree with comments about the output of the read_evalute_report() function. However, the function description/example does present @laijasmine's suggestion to use writeLines() and also mention the possibility of alternative output formats with the frmt argument (defaults to xml). I also noted that the read_evalute_report_summary() offers a compact way to visualise the report output, which is quite neat.

A new section on "Interpreting the Evaluation Report" has been added to the Evaluate and Upload Data vignette (see EDIorg/EDIutils@f4a0ab9). This provides more information on the report structure and content.

#### Retrieve Download and Citation metrics
>These vignettes are well structure and include reasonable examples and potential user cases. Functions and arguments are descriptive and detailed. No errors or issues to report for those vignettes, with exception of the citation doi error in the examples which were already mentioned by @laijasmine.

Thanks! The fix for the citation doi error is at EDIorg/EDIutils@743785c.
