### 2022-03-11 #498 @laijasmine
The description info is correct for me!

A couple minor revisions before I can approve. Now that some of the functions use the argument `as` we need to update the vignettes and Readme.

1) For this [vignette](https://ediorg.github.io/EDIutils/articles/retrieve_downloads.html), the default for `get_audit_report` has now changed to `"data.frame"` so we will need to update the code to specify `as` to be `xml`. Same goes for`list_data_package_citations` in this [vignette](https://ediorg.github.io/EDIutils/articles/retrieve_citations.html).

2) In both the [readme and the vignette](https://github.com/EDIorg/EDIutils/search?q=frmt) for `read_evaluate_report` the argument needs to be changed from `frmt` to `as`.

I think it is reasonable to leave the bigger changes (https://github.com/EDIorg/EDIutils/issues/34 and https://github.com/EDIorg/EDIutils/issues/35) as issues to be worked on in the future.

All the other changes looks good to me. Thank you for getting all those changes in!
