### 2022-03-14 40cee77
Fix function calls in the docs
The "frmt" parameter has been replaced by "as".
