+++ README.Rmd
@@ -159,7 +159,7 @@ status
 #> [1] TRUE
 
 # Read the evaluation report
-report <- read_evaluate_report(transaction, frmt = "char", env = "staging")
+report <- read_evaluate_report(transaction, as = "char", env = "staging")
 message(report)
 #> ===================================================
 #>   EVALUATION REPORT
