### 2022-03-22 #498 @bozaah
Thanks for following up Adam. I am satisfied with the replies and edits to my points.

Thank you for addressing our proposed changes @clnsmth. Also, I agree with @laijasmine, the bigger changes can be worked out in the future as issues.
