### 2022-03-23 deefe29
updated vignettes using the returned data frame
