+++ vignettes/retrieve_downloads.Rmd
@@ -11,18 +11,17 @@ This vignette demonstrates how to query the EDI repository for data package down
 
 ## Introduction to Download reports
 
-The PASTA Audit service collects information about user-related events that occur in PASTA, including data entity downloads. This information is stored in the Audit Manager's event database and which can be accessed through the `get_audit_report` command and returns an XML document composed of individual event audit records. Understanding this information is critical to determine the meaning of these events.
+The PASTA Audit service collects information about user-related events that occur in PASTA, including data entity downloads. This information is stored in the Audit Manager's event database and which can be accessed through the `get_audit_report` command and returns a data frame composed of individual event audit records. Understanding this information is critical to determine the meaning of these events.
 
 Attention needs to be paid to the `user` field, which can be `public` or `robot`. It is also important to know that the `userAgent` field is not required and a user agent can define its own text string to insert or leave blank. However, the `userAgent` provides a lot of interesting information. E.g., it allows a good estimate of downloads being initiated manually through a web browser (e.g., records starting with `Mozilla`) and downloads initiated programmatically (e.g., MATLAB, RStudio, etc.). The user agent `DataONE`, however, is the exception and appears to generate download reports that currently cannot be traced to actual, user generated entity downloads. They should be carefully inspected and cross checked with downloads reported for datasets on the DataONE search interface or filtered out as they seem to include internal management related data access. 
 
-Audit reports query a large database containing millions of records and may take some time to generate. In addition, a report containing a lot of audit records will create a large XML output, which may be slow to completely download and locally manage. It is best to use query parameters to limit requests.
+Audit reports query a large database containing millions of records and may take some time to generate. In addition, a report containing a lot of audit records will create a large output, which may be slow to completely download and locally manage. It is best to use query parameters to limit requests.
 
 ## Set up the R environment
 
 
 ```r
 library(EDIutils)
-library(xml2)
 library(stringr)
 library(dplyr)
 library(tidyr)
@@ -30,8 +29,6 @@ library(ggplot2)
 library(lubridate)
 ```
 
-
-Several EDI API functions return XML, `xml2` is a convenient package to extract desired information and to save returned EML files locally.
 
 ## Get the Audit Report for Downloaded Data Entities
 
@@ -66,50 +63,9 @@ query <- paste("category=info",
                sep = "&")
 
 # Get the report
-report_xml <- get_audit_report(query = query, env = "production")
+df_report <- get_audit_report(query = query, as = 'data.frame', env = "production")
 
 logout()
-```
-
-### Parse the Report
-
-The returned XML has the format:
-
-```xml
-<auditRecord>
-    <oid>115951546</oid>
-    <entryTime>2021-10-18T09:27:52</entryTime>
-    <category>info</category>
-    <service>DataPackageManager-1.0</service>
-    <serviceMethod>readDataEntity</serviceMethod>
-    <responseStatus>200</responseStatus>
-    <resourceId>https://pasta.lternet.edu/package/data/eml/knb-lter-ntl/387/1/c472d5a7ac8a89a12100792e9b5705d3</resourceId>
[…]
