+++ vignettes/retrieve_citations.Rmd
@@ -9,64 +9,83 @@ vignette: >
 
 This vignette demonstrates how to query the EDI repository for journal articles that cite a certain dataset. These stats may be used in reports.
 
+
 ```r
 library(EDIutils)
-library(xml2)
+library(dplyr)
+library(tidyr)
 ```
 
 ## Get the Journal Citation Report for Specific Data Packages
 
 Get all identifiers for a certain scope
 
+
 ```r
 scope <- "edi"
 identifiers <- list_data_package_identifiers(scope = scope, env = "production")
 ```
 
-Get all revisions for each identifer and retrieve journal citations for each. Citation information is connected to the specific revision of a dataset.
+Get the latest revisions for each identifer and retrieve all journal citations.
 
-The citation information is returned as XMl in the format:
-
-```
-<?xml version="1.0" encoding="UTF-8"?>
-<journalCitations>
-    <journalCitation>
-        <journalCitationId>15</journalCitationId>
-        <principalOwner>uid=ucarroll,o=LTER,dc=ecoinformatics,dc=org</principalOwner>
-        <dateCreated>2017-12-21T14:28:26.235</dateCreated>
-        <packageId>edi.1000.1</packageId>
-        <articleDoi>10.5072/FK2/06dccc7b0cb2a2d5f6fef62cb4b36dae</articleDoi>
-        <articleTitle>Tree Survey in Southern Arizona</articleTitle>
-        <articleUrl>http://treejournal.com/articles/12345</articleUrl>
-        <journalTitle>The Tree Journal</journalTitle>
-    </journalCitation>
-<journalCitations>
-```
+The citation information is returned as a data frame. First set up the empty data frame to append all records to.
 
 
 ```r
-citation_dois <- c()
+df_return_all <- data.frame(journalCitationId = character(0),
+                            packageId = character(0),
+                            principalOwner = character(0),
+                            dateCreated = character(0),
+                            articleDoi = character(0),
+                            articleTitle = character(0),
+                            articleUrl = character(0),
+                            journalTitle = character(0),
+                            relationType = character(0))
+
 for (j in 1:length(identifiers)){
-  revisions <- list_data_package_revisions(
+  revision <- list_data_package_revisions(
     scope = scope,
     identifier = identifiers[j],
-    env = "production"
-  )
+    filter = 'newest',
+    env = 'production')
+
+  datapackage <- paste(scope, identifiers[j], revision, sep = ".")
+  df_return <- list_data_package_citations(packageId = datapackage,
+                                           as = 'data.frame',
+                                           list_all = TRUE,
+                                           env = 'production')
   
-  for (i in 1:length(revisions)) {
-    datapackage <- paste(scope, identifiers[j],revisions[i], sep = ".")
-    return_xml <- list_data_package_citations(packageId = datapackage)
-    return_dois <- xml_text(xml_find_all(return_xml, "//articleDoi"))
-    citation_dois <- append(citation_dois, return_dois)
-  }
+  df_return_all <- rbind(df_return_all, df_return)
+
 }
 ```
 
-The list of `citation_dois` now contains all journal publication DOIs that have cited a dataset in the EDI scope (on 11/5/2021). This is not the number of journal articles because one article may cite more than one dataset. To get to the number of journal articles, the number of unique DOIs needs to be counted.
+Various analyses may now be conducted on this data frame. E.g., number of articles citing datasets 
 
 
 ```r
-print(length(citation_dois))
-#> [1] 272
+df_datasets <- df_return_all %>%
[…]
