### 2022-03-24 EDIorg/EDIutils#37 @clnsmth
xml2df() drops null elements
The internal function `xml2df()` drops null elements from the returned data.frame. The expected result is a data.frame with null elements filled with `NA`. This issue affects several EDItuils functions. See:

https://github.com/EDIorg/EDIutils/blob/5ee913a63c4127938ddadcf7bdbfbd622796fdf3/R/utilities.R#L426
