### 2022-03-24 #498 @clnsmth
@adamhsparks @laijasmine @bozaah A bug was discovered when rebuilding the two vignettes "Retrieve Download Metrics" and "Retrieve Citation Metrics" (https://github.com/EDIorg/EDIutils/issues/37). I'll send an update once this issue is fixed.
