### 2022-03-24 EDIorg/EDIutils#37 @clnsmth
Turns out this is an issue in one of the PASTA REST API methods, not `xml2df()`. See: https://github.com/PASTAplus/PASTA/issues/76

A patch can be applied to EDIutils or a global fix at the repository level.
