### 2022-03-25 5d27bc8
Update download and citation vignettes
* updated download metrics vignette.

* updated vignettes using the returned data frame
