+++ vignettes/retrieve_downloads.Rmd
@@ -7,44 +7,38 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-This vignette demonstrates how to query the EDI repository for data package download metrics. These stats may be used in reports or further processed to understand data user behavior.
+This vignette demonstrates how to query the EDI repository for data package download metrics. These stats may be used in reports or further processed to understand data user behavior. Although EDI makes every effort of accurate reporting, these download data need to be carefully inspected and understood before using.
+
+## Introduction to Download reports
+
+The PASTA Audit service collects information about user-related events that occur in PASTA, including data entity downloads. This information is stored in the Audit Manager's event database and which can be accessed through the `get_audit_report` command and returns a data frame composed of individual event audit records. Understanding this information is critical to determine the meaning of these events.
+
+Attention needs to be paid to the `user` field, which can be `public` or `robot`. It is also important to know that the `userAgent` field is not required and a user agent can define its own text string to insert or leave blank. However, the `userAgent` provides a lot of interesting information. E.g., it allows a good estimate of downloads being initiated manually through a web browser (e.g., records starting with `Mozilla`) and downloads initiated programmatically (e.g., MATLAB, RStudio, etc.). The user agent `DataONE`, however, is the exception and appears to generate download reports that currently cannot be traced to actual, user generated entity downloads. They should be carefully inspected and cross checked with downloads reported for datasets on the DataONE search interface or filtered out as they seem to include internal management related data access. 
+
+Audit reports query a large database containing millions of records and may take some time to generate. In addition, a report containing a lot of audit records will create a large output, which may be slow to completely download and locally manage. It is best to use query parameters to limit requests.
 
 ## Set up the R environment
 
+
 ```r
 library(EDIutils)
-library(xml2)
 library(stringr)
 library(dplyr)
-#> 
-#> Attaching package: 'dplyr'
-#> The following objects are masked from 'package:stats':
-#> 
-#>     filter, lag
-#> The following objects are masked from 'package:base':
-#> 
-#>     intersect, setdiff, setequal, union
-library(lubridate)
-#> 
-#> Attaching package: 'lubridate'
-#> The following objects are masked from 'package:base':
-#> 
-#>     date, intersect, setdiff, union
+library(tidyr)
 library(ggplot2)
+library(lubridate)
 ```
 
-Several EDI API functions return XML, `xml2` is a convenient package to extract desired information and to save returned EML files locally.
 
 ## Get the Audit Report for Downloaded Data Entities
 
 Accessing the audit report requires authentication.
 
-In a text editor create the file `.Renviron` with the content:
 
-```{r eval=FALSE}
-login()
-#> User name: "my_name"
-#> User password: "my_secret"
+```r
+login(userId = '', userPass = '')
+#> userID: "my_name"
+#> userPass: "my_secret"
 ```
 
 Setting the query parameters to get the desired audit report:
@@ -58,7 +52,8 @@ Setting the query parameters to get the desired audit report:
   
 For more information on searchable fields see `get_audit_report()` documentation.
   
-```{r eval=FALSE}
+
+```r
 # Construct the query
 query <- paste("category=info",
[…]
