### 2022-03-25 9e52ab4
Lint download and citation vignettes
