### 2022-03-26 61204da
Use canonical URL for CRAN packages
