+++ R/read_metadata.R
@@ -7,9 +7,9 @@
 #' @return (xml_document) EML metadata document.
 #' 
 #' See the 
-#' \href{https://cran.r-project.org/web/packages/emld/index.html}{emld} library 
+#' \href{https://CRAN.R-project.org/package=emld}{emld} library 
 #' for more on working with EML as a list or JSON-LD. See the 
-#' \href{https://cran.r-project.org/web/packages/xml2/index.html}{xml2} library 
+#' \href{https://CRAN.R-project.org/package=xml2}{xml2} library 
 #' for working with EML as XML.
 #' 
 #' @family Accessing
