+++ man/get_event_subscription_schema.Rd
@@ -15,7 +15,7 @@ get_event_subscription_schema(env = "production")
 entities.
 
 See the 
-\href{https://cran.r-project.org/web/packages/xml2/index.html}{xml2} library 
+\href{https://CRAN.R-project.org/package=xml2}{xml2} library 
 for more on working with XML.
 }
 \description{
