+++ man/read_metadata_dublin_core.Rd
@@ -16,7 +16,7 @@ read_metadata_dublin_core(packageId, env = "production")
 (xml_document) Dublin Core metadata.
 
 See the 
-\href{https://cran.r-project.org/web/packages/xml2/index.html}{xml2} library 
+\href{https://CRAN.R-project.org/package=xml2}{xml2} library 
 for more on working with XML.
 }
 \description{
