### 2022-03-27 #498 @clnsmth
@adamhsparks, @laijasmine, @bozaah
- The bug is fixed (see EDIorg/EDIutils#37)
- @laijasmine 's requested changes (https://github.com/ropensci/software-review/issues/498#issuecomment-1064787189) are are implemented
- version 0.0.0.9003 has been tested on R-hub with results listed in [cran-comments.md](https://github.com/EDIorg/EDIutils/blob/main/cran-comments.md)
- All above changes are in the `main` branch of EDIutils
