+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: EDIutils
 Title: An API Client for the Environmental Data Initiative Repository
-Version: 0.0.0.9003
+Version: 0.0.0.9004
 Authors@R: 
     c(person("Colin", "Smith", email = "colin.smith@wisc.edu", role = c("aut", "cre"), comment = "https://orcid.org/0000-0003-2261-9931"),
     person("Corinna", "Gries", role = "ctb", comment = "https://orcid.org/0000-0002-9091-6543"),
