### 2022-04-21 #498 @maelle
Note that we now require 2 factor authentication for organization members, see https://docs.github.com/en/authentication/securing-your-account-with-two-factor-authentication-2fa/configuring-two-factor-authentication so you'll have to enable it before joining the organization (in case you haven't enabled it yet).
